cv-intake-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
httpx>=0.24.1
redis>=4.5.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
boto3>=1.37.0
tenacity>=8.2.0
pyjwt>=2.8.0
slowapi>=0.1.9
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
prometheus-client>=0.17.0